+++ wpull/app.py
@@ -51,6 +51,7 @@ class Application(HookableMixin):
         self._event_loop = trollius.get_event_loop()
         self._exit_code = 0
         self._statistics = None
+        self.stop_observer = wpull.observer.Observer()
 
         self.register_hook('exit_status', 'finishing_statistics')
 
@@ -137,6 +138,7 @@ class Application(HookableMixin):
             pass
 
         self._print_stats()
+        self.stop_observer.notify()
 
         raise tornado.gen.Return(self._exit_code)